+++ src/lib/pages/all-codes/index.tsx
@@ -34,7 +34,7 @@ const AllCodes = observer(() => {
         </Heading>
         <Flex gap={2}>
           <InputWithIcon
-            placeholder="Search with code ID or code description"
+            placeholder="Search with code ID or code name"
             value={keyword}
             onChange={(e: ChangeEvent<HTMLInputElement>) =>
               setValue("keyword", e.target.value)