+++ src/lib/utils/description.ts
@@ -1,2 +1,0 @@
-export const getDescriptionDefault = (description?: string) =>
-  description ?? "";