+++ src/lib/utils/descriptionAndName.ts
@@ -1,0 +1,1 @@
+export const getNameAndDescriptionDefault = (value?: string) => value ?? "";